+++ app/src/main/java/cn/wthee/pcrtool/data/enums/AttrValueType.kt
@@ -1,0 +1,10 @@
+package cn.wthee.pcrtool.data.enums
+
+/**
+ * 属性值类型
+ */
+enum class AttrValueType {
+    INT,
+    DOUBLE,
+    PERCENT
+}
