+++ app/src/main/java/cn/wthee/pcrtool/data/enums/IconResourceType.kt
@@ -1,0 +1,12 @@
+package cn.wthee.pcrtool.data.enums;
+
+/**
+ * 图标类型
+ */
+enum class IconResourceType {
+    CHARACTER,
+    EQUIP,
+    UNIQUE_EQUIP,
+    EX_EQUIP,
+    ;
+}
