+++ app/src/main/java/cn/wthee/pcrtool/data/enums/KeywordType.kt
@@ -1,0 +1,9 @@
+package cn.wthee.pcrtool.data.enums;
+
+/**
+ * 查询关键词
+ */
+enum class KeywordType(val type: Int) {
+    NEWS(1),
+    TWEET(2);
+}
