+++ app/src/main/java/cn/wthee/pcrtool/data/enums/NewsType.kt
@@ -1,0 +1,15 @@
+package cn.wthee.pcrtool.data.enums;
+
+import cn.wthee.pcrtool.R
+
+/**
+ * 公告
+ */
+enum class NewsType(val stringId: Int) {
+    UPDATE(R.string.update),
+    SYSTEM(R.string.system),
+    NEWS(R.string.tool_news),
+    EVENT(R.string.event),
+    SHOP(R.string.shop),
+    LOCAL(R.string.local_note);
+}
