+++ app/src/main/java/cn/wthee/pcrtool/data/enums/SettingSwitchType.kt
@@ -1,0 +1,26 @@
+package cn.wthee.pcrtool.data.enums
+
+/**
+ * 设置枚举
+ */
+enum class SettingSwitchType {
+    /**
+     * 振动设置
+     */
+    VIBRATE,
+
+    /**
+     * 动画效果设置
+     */
+    ANIMATION,
+
+    /**
+     * 动态色彩设置
+     */
+    DYNAMIC_COLOR,
+
+    /**
+     * 使用ip访问（仅网络异常使用）
+     */
+    USE_IP
+}
