+++ app/src/main/java/cn/wthee/pcrtool/data/enums/SkillIndexType.kt
@@ -1,0 +1,41 @@
+package cn.wthee.pcrtool.data.enums;
+
+/**
+ * 技能下标类型
+ */
+enum class SkillIndexType(val index: Int) {
+    UNKNOWN(-1),
+    UB(0),
+    UB_PLUS(0),
+    MAIN_SKILL_1(1),
+    MAIN_SKILL_2(2),
+    MAIN_SKILL_3(3),
+    MAIN_SKILL_4(4),
+    MAIN_SKILL_5(5),
+    MAIN_SKILL_6(6),
+    MAIN_SKILL_7(7),
+    MAIN_SKILL_8(8),
+    MAIN_SKILL_9(9),
+    MAIN_SKILL_10(10),
+    MAIN_SKILL_1_PLUS(1),
+    MAIN_SKILL_2_PLUS(2),
+    EX_1(1),
+    EX_2(2),
+    EX_3(3),
+    EX_4(4),
+    EX_5(5),
+    EX_1_PLUS(1),
+    EX_2_PLUS(2),
+    EX_3_PLUS(3),
+    EX_4_PLUS(4),
+    EX_5_PLUS(5),
+    SP_UB(0),
+    SP_SKILL_1(1),
+    SP_SKILL_2(2),
+    SP_SKILL_3(3),
+    SP_SKILL_4(4),
+    SP_SKILL_5(5),
+    SP_SKILL_1_PLUS(1),
+    SP_SKILL_2_PLUS(2),
+    ;
+}
