+++ app/src/main/java/cn/wthee/pcrtool/data/enums/SkillType.kt
@@ -1,0 +1,10 @@
+package cn.wthee.pcrtool.data.enums
+
+/**
+ * 技能类型
+ */
+enum class SkillType {
+    ALL,
+    NORMAL,
+    SP
+}
